chore: rename to avoid "game"
